+++ backend/clouddm-platform/cgdm-components/cg-com-results-file/src/main/java/com/clougence/clouddm/component/resultfile/ResultFileWriter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.clougence.clouddm.platform.component.resultfile;
+package com.clougence.clouddm.component.resultfile;
 
 import java.io.Closeable;
 import java.io.File;
