+++ backend/clouddm-platform/cgdm-components/cg-com-results-file/src/main/java/com/clougence/clouddm/component/resultfile/ResultSetOverflowException.java
@@ -13,10 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.clougence.clouddm.platform.component.resultfile;
+package com.clougence.clouddm.component.resultfile;
 
-public interface ResultDataTag {
+import java.io.IOException;
 
-    int DATA_CROPPED_TAG = 0x01;
-    int DATA_MASK_TAG    = 0x02;
+import lombok.Getter;
+
+@Getter
+public class ResultSetOverflowException extends IOException {
+
+    private final long overflowSize;
+
+    public ResultSetOverflowException(String s, long overflowSize){
+        super(s);
+        this.overflowSize = overflowSize;
+    }
 }
